refactor: create new "src/plonk/circuit/helpers.rs" file
